refactor: `LinesSection` as its own component
